Create README.md
The goal of this project was to build a machine learning model capable of accurately predicting depression in a population where the incidence rate is 15%. The challenge was to prioritize recall—ensuring that the model captures as many true positive cases as possible—while also maintaining reasonable precision to limit false positives, given budgetary constraints. Using the XGBoost algorithm, we employed a structured approach that included data preparation, hyperparameter tuning, and threshold classification manipulation to optimize model performance. The techniques applied, particularly hyperparameter tuning and adjusting the classification threshold, were highly effective, ultimately leading to a substantial improvement in the F1 score and meeting our project goals.